+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- allow processing of phony recipes with no requirements
+
 ## [0.1.0] - 2020-01-30
 ### Added
 - add the yamk command as a placeholder